+++ components/genApp/Layouts.js
@@ -1,6 +1,6 @@
 import React from 'react'
 // import Dashboard from './Dashboard'
-import style from './styles/style'
+import style from '../styles/style'
 import Head from 'next/head'
 
 class Layouts extends React.Component {
