+++ components/profile/setUp/Rooms.js
@@ -1,0 +1,9 @@
+import Header from '../../genApp/Header'
+
+const Rooms = () => (
+  <div>
+    <Header />
+  </div>
+)
+
+export default Rooms
