--- pages/room-profile.js
@@ -1,8 +1,0 @@
-import Layout from '../components/Layouts.js'
-
-export default (props) => (
-  <Layout>
-    <h1>{props.url.query.title}</h1>
-    <p>This is the blog post content.</p>
-  </Layout>
-)
