+++ pages/setup/rooms.js
@@ -1,0 +1,12 @@
+import Layout from '../../components/genApp/Layouts'
+import Rooms from '../../components/profile/setUp/Rooms'
+
+const RoomProfile = (props) => (
+  <div>
+    <Layout>
+      <Rooms />
+    </Layout>
+  </div>
+)
+
+export default RoomProfile
